+++ ShakerSort.h
@@ -48,9 +48,9 @@ private:
     void saveArrayToFile(int [], const std::string&) const;
     void readArrayFromFile();
     static std::string getFileName();
-    void printExecutionTime(double &);
+    void printExecutionTime(long &);
     static void printDashedLine();
-    static std::string getExecutionTime(double &);
+    static std::string getExecutionTime(long &);
     /// member variables and functions which are public
 public:
     /**
